requirements.txt
Django~=4.2.5
djangorestframework~=3.17.1
dj-rest-auth~=7.2.0
django-allauth~=65.16.1
django-cors-headers~=4.9.0
psycopg2-binary~=2.9.6
gunicorn~=20.1.0
python-dotenv~=0.21.0
pillow~=12.2.0
python-dotenv~=1.2.2